+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - added `change-grayscale` filter for change pixel values of grayscale images by a factor of fixed value
 - added `clip-grayscale` filter for replacing pixel values that go below or above thresholds
+- added `enhance` filter for enhancing brightness, contrast, color, sharpness
 
 
 0.0.1 (2024-05-06)